Actualización del backend y entidades nuevas

+++ src/main/java/consultorio/modelo/Paciente.java
@@ -9,7 +9,7 @@ import java.time.LocalDate;
 public class Paciente extends BaseEntidad {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private String nombre;
     private String apellido;
@@ -39,8 +39,15 @@ public class Paciente extends BaseEntidad {
     }
 
     // Getters y setters
-    public Integer getId() { return id; }
-    public void setId(Integer id) { this.id = id; }
+
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
 
     public String getNombre() { return nombre; }
     public void setNombre(String nombre) { this.nombre = nombre; }
